Remove v3 code
