+++ src/commonMain/kotlin/exchange.dydx.abacus/protocols/PublicProtocols.kt
@@ -20,8 +20,8 @@ import exchange.dydx.abacus.output.TransferStatus
 import exchange.dydx.abacus.output.Wallet
 import exchange.dydx.abacus.output.input.Input
 import exchange.dydx.abacus.responses.ParsingError
-import exchange.dydx.abacus.state.app.ApiState
 import exchange.dydx.abacus.state.changes.StateChanges
+import exchange.dydx.abacus.state.manager.ApiState
 import exchange.dydx.abacus.utils.IList
 import exchange.dydx.abacus.utils.IMap
 import kollections.JsExport
