+++ src/commonMain/kotlin/exchange.dydx.abacus/state/manager/V4StateManagerAdaptor.kt
@@ -11,16 +11,7 @@ import exchange.dydx.abacus.protocols.TransactionCallback
 import exchange.dydx.abacus.protocols.TransactionType
 import exchange.dydx.abacus.responses.ParsingError
 import exchange.dydx.abacus.responses.ParsingErrorType
-import exchange.dydx.abacus.responses.StateResponse
-import exchange.dydx.abacus.state.app.ApiState
-import exchange.dydx.abacus.state.app.ApiStatus
-import exchange.dydx.abacus.state.app.IndexerURIs
-import exchange.dydx.abacus.state.app.NetworkState
-import exchange.dydx.abacus.state.app.NetworkStatus
-import exchange.dydx.abacus.state.app.V4Environment
 import exchange.dydx.abacus.state.app.adaptors.V4TransactionErrors
-import exchange.dydx.abacus.state.changes.Changes
-import exchange.dydx.abacus.state.changes.StateChanges
 import exchange.dydx.abacus.state.manager.configs.V4StateManagerConfigs
 import exchange.dydx.abacus.state.modal.TransferInputField
 import exchange.dydx.abacus.state.modal.onChainAccountBalances
@@ -33,16 +24,12 @@ import exchange.dydx.abacus.state.modal.onChainUserFeeTier
 import exchange.dydx.abacus.state.modal.onChainUserStats
 import exchange.dydx.abacus.state.modal.squidChains
 import exchange.dydx.abacus.state.modal.squidTokens
-import exchange.dydx.abacus.state.modal.transfer
 import exchange.dydx.abacus.utils.CoroutineTimer
 import exchange.dydx.abacus.utils.IMap
 import exchange.dydx.abacus.utils.IOImplementations
 import exchange.dydx.abacus.utils.UIImplementations
 import exchange.dydx.abacus.utils.iMapOf
 import exchange.dydx.abacus.utils.isAddressValid
-import io.ktor.client.utils.EmptyContent.status
-import kollections.iListOf
-import kollections.iMapOf
 import kollections.toIMap
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
