+++ src/commonMain/kotlin/exchange.dydx.abacus/state/manager/configs/StateManagerConfigs.kt
@@ -1,7 +1,7 @@
 package exchange.dydx.abacus.state.manager.configs
 
-import exchange.dydx.abacus.state.app.IndexerURIs
-import exchange.dydx.abacus.state.app.V4Environment
+import exchange.dydx.abacus.state.manager.IndexerURIs
+import exchange.dydx.abacus.state.manager.V4Environment
 import exchange.dydx.abacus.utils.IList
 import exchange.dydx.abacus.utils.IMap
 import exchange.dydx.abacus.utils.Parser
